LITE-22101 Django-RQL now relies on Lib-RQL
