+++ .github/workflows/build.yml
@@ -54,12 +54,6 @@ jobs:
     - name: Testing
       run: |
         python setup.py test
-    - name: Upload coverage
-      uses: codecov/codecov-action@v1
-      with:
-        file: ./tests/reports/coverage.xml
-        fail_ci_if_error: true
-        verbose: true
     - name: Fix coverage report for Sonar
       run: |
         sed -i 's/\/home\/runner\/work\/django-rql\/django-rql\//\/github\/workspace\//g' ./tests/reports/coverage.xml
