+++ requirements/dev.txt
@@ -1,3 +1,2 @@
-lark-parser==0.11.0
+lib-rql
 Django>=2.2.19
-cachetools>=4.2.4
